Create The Set of Intelligence (Cardiovascular System).md
This project replicates the work of the Hodgkin and Huxley model to characterize the long-lasting action and pace-maker potentials of the Purkinje fibers of the heart, with some adjustments. It describes the computations of the electrical Conduction in the Heart coordinates contraction. Also, it talks about the cardiac conduction system, and how the electrical signal travels from the SA node (the pacemaker of the heart) through internodal pathways to the AV node. Moreover, how Purkinje fibers play a major role in electrical conduction and propagation of impulse to the ventricular muscle, and that it allows the heart's conduction system to create synchronized contractions of its ventricles and are essential for maintaining a consistent heart rhythm. In addition, the action potential and the four phases, from phase 4 (resting phase) to phase 3 (repolarization). Moreover, the potassium and sodium current in the Hodgkin and Huxley model along with the results of the mathematical model.